Do not return error in lambda when error is not recognized by the app
Why: The rationale to return this types of errors originally was so that Lambda retried the invocation, however if all retries failed this provoked APIGateway to return 502 instead of 500. This error is not understandable for a regular user. So I decided to remove this functionality to test whether is the best approach.