Pass user object to the method that creates users in the DB
Why: To create the default categories for a user, there is some logic involved related to the generation of category IDs. This logic lived on the repository so I decided to move it to the usecases. In order for that to work I had to pass the user object to the creation method.