Do not use array of pointers to accumulate queried items
Why: I was using two array of pointers: one to accumulate the queried items(expensesEntites), another to contain the items in the current query(itemsInQuery). Given the fact that the queried items were pointers, each time a new query was made, the items in expensesEntites also changed. Both arrays pointed to literally the same items. This meant that expensesEntites was not accumulating items.