[router] Use regular logger in router and pass it as param to the lambdas
Why: So that the router can also log to the Logstash server. Previously, I had decided to open two connections to the Logstash server: one for the lambdas and another for the router. This was done in goroutines, so sometimes when closing one of the connections, a race condition occurred, causing the error "error closing connection to Logstash server: close tcp 169.254.76.1:49854->54.234.100.95:5044: use of closed network connection". This commit is the solution to that issue.