[logger] Make logger connection nil after closing it
Why: The code makes a nil check on the connection to know whether the connection is already closed. If I don't explicitly set the
connection field to nil, another function may try to close an already closed connection, provoking an error. As it turns out, the net.Close() method doesn't make the connection nil.