[frontend] Use axios package directly instead of axios instance for logout
Why: This api call can be made AFTER a 401 response from the server. Given that the axios instance has an interceptor for 401 statuses, using it for logout may produce an infinite loop, which is the reason why I suspect the /logout request is never made when using the axios instance.