[users] Return 500 when user is not found for GET /users endpoint
Why: The username that’s used to get the user from the DB is the one on the tokens that are emitted by the server. Before the server creates the tokens, it supposed to have persisted the user on the DB, so if the user can’t be found, then the server did something wrong.