[expenses repository] Create method to update expenses in batches
Why: Set periods to expenses that do not have it.
Note: As of right now, the implementation of this function is the same
as the one that creates expenses in batches because the operation to
create and update items in Dynamo is the same(PutItem). Despite that, I
decided to create a separate function to update expenses to avoid
confusion and also to accommodate possible future variations.