[expenses] Make an expense's period a string value instead of pointer value
Why: Using pointers as fields of a model introduces certain amount of complexity that only is worth it if a "nil" value has a different meaning that a type's empty value in the context of the application. In the case of an expense's period, "nil" means the same as en empty string, so it doesn't make sense for this to be a pointer.