[storage, period-stat] Move period stat repo to "expenses" repo directory
Why: Despite it's name, a "period-stat" entity is much more related to
an expenses than to a period. The only way a period-stat is created is
when an expense is. Having those repos in separate packages made the
task of creating a period-stat in the same transaction as an expense, a
lot more difficult.