[savings api] Check if the nextKey is empty to break the infinite loop that calculates the progress by saving goal
Why: The loop was never finishing because I only checked for errors to
break it. However, there is a case in which the next key is empty and
there is no error returned from the repo. Remember that *sometimes*
Dynamo knows when it has retrieved the last item. Its not always
necessary to query the next page to know this.