[ExpensesTable] Show Expenses error snackbar only on non-404 errors
Why: This type of alert is too alarming for a not found error in this
case. Is we have no expenses it might be because the user hasn't create
any yet, or maybe because the filters applied in the table don't yield
any results. In any case, the "No expenses found" message that appears
in the table is enough